Move NativeMethods to System namespace
